test: have ready dockerfile
