+++ crafters/nlp/SpellChecker/Dockerfile
@@ -1,4 +1,4 @@
-FROM jinaai/jina:1.1.5
+FROM jinaai/jina:1.1.5 as base
 
 # setup the workspace
 COPY . /workspace
@@ -8,6 +8,9 @@ WORKDIR /workspace
 RUN pip install -r requirements.txt
 
 # for testing the image
+FROM base
 RUN pip install pytest && pytest -v -s
 
+FROM base
+RUN python pyngramspell/train.py
 ENTRYPOINT ["jina", "pod", "--uses", "config.yml"]
